+++ modules/terracotta-provider-hangar/src/main/kotlin/io/github/beduality/terracotta/provider/hangar/model/HangarChannel.kt
@@ -1,0 +1,9 @@
+package io.github.beduality.terracotta.provider.hangar.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class HangarChannel(
+    val name: String,
+    val color: String = "green",
+)